requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
datasets>=2.12.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
opencv-python>=4.8.0
pillow>=10.0.0
rasterio>=1.3.0
geopandas>=0.13.0
folium>=0.14.0
requests>=2.31.0
beautifulsoup4>=4.12.0
tweepy>=4.14.0
yfinance>=0.2.0
newsapi-python>=0.2.6
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
alembic>=1.11.0
redis>=4.6.0
celery>=5.3.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
streamlit>=1.25.0
jupyter>=1.0.0
ipywidgets>=8.0.0
tqdm>=4.65.0
wandb>=0.15.0
mlflow>=2.5.0
hydra-core>=1.3.0
omegaconf>=2.3.0
boto3>=1.28.0
google-cloud-storage>=2.10.0
